Drop ten no-op (s8) casts on already-signed reads
gNrarewareStage / gNrarewareExitDelay (engine/51), gSaveSelectPanelIndex
(53), gOptionsActivePanel (55), gSaveGameCurrentSlot (23) and
ObjSequence.slot (441) are all declared s8 already, so the casts on their
reads add nothing. All five objects byte-identical.

+++ src/dlls/engine/23/23.c
@@ -354,7 +354,7 @@ void gplaySaveGame(int param)
             memcpy((void*)pRestartPoint, gSaveGameData, 0x564);
         }
     }
-    if ((s8)gSaveGameCurrentSlot == -1)
+    if (gSaveGameCurrentSlot == -1)
     {
         gSaveGameCurrentSlot = 0;
     }
@@ -387,7 +387,7 @@ void saveGame_save(void)
             memcpy((void*)pRestartPoint, gSaveGameData, 0x564);
         }
     }
-    if ((s8)gSaveGameCurrentSlot == -1)
+    if (gSaveGameCurrentSlot == -1)
     {
         gSaveGameCurrentSlot = 0;
     }
